C64.emu updates and misc tweaks
* Imagine: Fix if/else logic in GLRenderer::checkExtensionString()
* EmuFramework: Don't implicitly call showEmulation() in EmuApp::reloadSystem()
* C64.emu: Update core to VICE 3.8
* C64.emu: Add color settings in Options -> Video
* C64.emu: Save options per core: default TDE, border mode, crop normal borders, SID engine, ReSID resampling
* C64.emu: Add "Just Restart" to quick settings menu
* C64.emu: Add C64System::enterCPUTrap() to safely modify various settings outside of vblank
* C64.emu: Convert all options to use the new stateless API to use the VICE resource API directly and avoid stale cached values
* C64.emu: Only call onFrameTimeChanged() in vsyncarch_refresh_frequency_changed() if content is loaded

+++ C64.emu/src/vice/c64/c64cia2.c
@@ -147,7 +147,7 @@ static void store_ciapa(cia_context_t *cia_context, CLOCK rclk, uint8_t byte)
             store_userport_pa3((uint8_t)((byte & 8) >> 3));
         }
 
-        tmp = ~byte;
+        tmp = (uint8_t)~byte;
         new_vbank = tmp & 3;
         if (new_vbank != vbank) {
             vbank = new_vbank;
