+++ C64.emu/src/vice/monitor/mon_assemble6809.c
@@ -245,7 +245,7 @@ static int mon_assemble_instr(const char *opcode_name, asm_mode_addr_info_t oper
     }
 
     for (i = 0; i < len; i++) {
-        mon_set_mem_val(mem, (uint8_t)(loc + i), opc[i]);
+        mon_set_mem_val(mem, (uint16_t)(loc + i), opc[i]);
     }
 
     if (len >= 0) {
